refactor: split validationUtils into separate validateName, validateEmail, validateAge folders with co-located tests
