+++ src/utils/validateAge/validateAge.ts
@@ -1,0 +1,1 @@
+export const validateAge = (age: number) => age > 0;
