+++ src/utils/validateEmail/validateEmail.ts
@@ -1,8 +1,3 @@
-export const validateName = (name: string) =>
-  name.length > 0 &&
-  name[0] === name[0].toUpperCase() &&
-  name[0] !== name[0].toLowerCase();
-
 export const validateEmail = (email: string) => {
   const parts = email.split('@');
   if (parts.length !== 2) return false;
@@ -11,5 +6,3 @@ export const validateEmail = (email: string) => {
   if (!domain.includes('.')) return false;
   return domain.split('.').every((p) => p !== '');
 };
-
-export const validateAge = (age: number) => age > 0;
